chore: move test utils under __tests__
